pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "korail2"
version = "0.4.0"
description = "Korail(www.letskorail.com) wrapper for Python"
readme = "README.rst"
license = { file = "LICENSE" }
authors = [{ name = "Taehoon Kim", email = "carpedm20@gmail.com" }]
requires-python = ">=3.8"
dependencies = [
    "requests",
    "six",
    "pycryptodome",
    "python-dotenv",
]

[project.urls]
Homepage = "http://github.com/carpedm20/korail2"
